project/requirements.txt
# Web server & framework
Flask==3.0.3
gunicorn==21.2.0

# Env vars
python-dotenv==1.0.1

# HTTP client
requests==2.32.3

# Google Sheets
gspread==6.1.2
google-auth==2.33.0

# OpenAI SDK (GPT-4o etc.)
openai==1.42.0

# Utils (opționale dar utile și folosite în codul propus)
dateparser==1.2.0      # pentru expresii gen "mâine", "în 3 zile", "10 septembrie"
rapidfuzz==3.9.6       # căutare fuzzy în catalog
tenacity==8.5.0        # retry cu backoff, dacă îl folosești în integrări
